FIX Forsøk på å fiksa feil i besluttar-panelet der aksjonspunkta av og til blir lagra i feil rekkefølge
